+++ lib/features/interactive.js
@@ -69,18 +69,19 @@ export default class WebsocketServer {
     debug(this.url);
 
     var queue = task.runtime.queue;
-    var expiration = new Date(Date.now() + 60 * 60 * 24);
+    debug('time %s',task.task.payload.maxRunTime);
+    var expiration = new Date(Date.now() + task.task.payload.maxRunTime * 1000);
     await queue.createArtifact(
       task.status.taskId,
       task.runId,
-      'interactive', {
+      'private/mozilla/interactive.sock', {
         storageType: 'reference',
         expires: expiration.toJSON(),
         contentType: 'application/octet-stream',
         url: this.url
       }
     );
-    
+    debug('artifact made');
   }
 
   async killed (task) {
